refactor project structure/ setup MVP for ProjectListActivity
