+++ WaveExample/app/src/main/AndroidManifest.xml
@@ -8,7 +8,7 @@
     android:label="@string/app_name"
     android:supportsRtl="true"
     android:theme="@style/AppTheme">
-    <activity android:name=".ProjectListActivity">
+    <activity android:name=".projectlist.ProjectListActivity">
       <intent-filter>
         <action android:name="android.intent.action.MAIN"/>
