--- WaveExample/app/src/main/kotlin/com/derekcsm/wave_example/ProjectListActivity.kt
@@ -1,12 +1,0 @@
-package com.derekcsm.wave_example
-
-import android.support.v7.app.AppCompatActivity
-import android.os.Bundle
-
-class ProjectListActivity : AppCompatActivity() {
-
-  override fun onCreate(savedInstanceState: Bundle?) {
-    super.onCreate(savedInstanceState)
-    setContentView(R.layout.act_project_list)
-  }
-}
