+++ WaveExample/app/src/main/kotlin/com/derekcsm/wave_example/projectlist/ProjectListContract.kt
@@ -1,0 +1,13 @@
+package com.derekcsm.wave_example.projectlist
+
+interface ProjectListContract {
+
+  interface View {
+
+  }
+
+  interface UserActionsListener {
+    fun test()
+  }
+
+}
