+++ WaveExample/app/src/main/kotlin/com/derekcsm/wave_example/projectlist/ProjectListPresenter.kt
@@ -1,0 +1,18 @@
+package com.derekcsm.wave_example.projectlist
+
+import android.util.Log
+
+class ProjectListPresenter constructor(mView: ProjectListContract.View) :
+    ProjectListContract.UserActionsListener {
+
+  private val TAG = "ProjectListPresenter"
+  private val mView: ProjectListContract.View
+
+  init {
+    this.mView = mView
+  }
+
+  override fun test() {
+    Log.d(TAG, "test")
+  }
+}
