+++ WaveExample/app/src/main/res/layout/act_project_list.xml
@@ -9,7 +9,7 @@
   android:paddingLeft="@dimen/activity_horizontal_margin"
   android:paddingRight="@dimen/activity_horizontal_margin"
   android:paddingTop="@dimen/activity_vertical_margin"
-  tools:context="com.derekcsm.wave_example.ProjectListActivity">
+  tools:context="com.derekcsm.wave_example.projectlist.ProjectListActivity">
 
   <TextView
     android:layout_width="wrap_content"
